folder restructuring: extract monolith files into its own folder instead of the root folder
