+++ monolith/index.js
@@ -4,7 +4,7 @@ const axios = require('axios');
 
 const typeDefs = gql(readFileSync('./schema.graphql', { encoding: 'utf-8' }));
 const resolvers = require('./resolvers');
-const { BookingsDataSource, ReviewsDataSource, ListingsAPI, AccountsAPI, PaymentsAPI } = require('./services');
+const { BookingsDataSource, ReviewsDataSource, ListingsAPI, AccountsAPI, PaymentsAPI } = require('../services');
 
 require('dotenv').config();
 
@@ -24,11 +24,9 @@ const server = new ApolloServer({
     const token = req.headers.authorization || '';
     const userId = token.split(' ')[1]; // get the user name after 'Bearer '
     if (userId) {
-      const { data } = await axios
-        .get(`http://localhost:4011/login/${userId}`)
-        .catch((error) => {
-          throw new AuthenticationError(error.message);
-        });
+      const { data } = await axios.get(`http://localhost:4011/login/${userId}`).catch((error) => {
+        throw new AuthenticationError(error.message);
+      });
 
       return { userId: data.id, userRole: data.role };
     }
