+++ services/index.js
@@ -1,8 +1,8 @@
-const BookingsDataSource = require('../datasources/bookings');
-const ReviewsDataSource = require('../datasources/reviews');
-const ListingsAPI = require('../datasources/listings');
-const AccountsAPI = require('../datasources/accounts');
-const PaymentsAPI = require('../datasources/payments');
+const BookingsDataSource = require('../monolith/datasources/bookings');
+const ReviewsDataSource = require('../monolith/datasources/reviews');
+const ListingsAPI = require('../monolith/datasources/listings');
+const AccountsAPI = require('../monolith/datasources/accounts');
+const PaymentsAPI = require('../monolith/datasources/payments');
 
 module.exports = {
   BookingsDataSource,
